+++ src/app/layout/main/main.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { FooterComponent } from '../footer/footer.component';
-import { Docs_T01_IntoComponent } from '../../docs/t01-starter/into-getting-starter/into-title-01.component';
+import { Docs_T01_IntoComponent } from '../../docs/t01-starter/into-t01-getting-starter/into-title-01.component';
 import { Docs_T02_IntoComponent } from '../../docs/t02-concepts/into-t02-main-concepts/into-title-02.component';
 import { RouterOutlet } from '@angular/router';
